requirements.txt
streamlit==1.40.2
streamlit-chat==0.1.1
langchain-community==0.3.8
sentence-transformers==3.3.1
pinecone-client==5.0.1
langchain-pinecone==0.2.0
langchain==0.3.9
openai==1.55.2